+++ src/test/java/com/epam/reportportal/service/ReportPortalTest.java
@@ -33,7 +33,6 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import org.apache.commons.lang3.tuple.Pair;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -43,10 +42,12 @@ import org.slf4j.LoggerFactory;
 import org.slf4j.bridge.SLF4JBridgeHandler;
 
 import javax.net.ssl.*;
+import java.io.InputStream;
 import java.net.InetAddress;
 import java.net.ServerSocket;
 import java.nio.charset.StandardCharsets;
 import java.security.KeyStore;
+import java.security.SecureRandom;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.*;
@@ -65,13 +66,7 @@ import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.*;
 
 public class ReportPortalTest {
-	private static final String TRUSTSTORE_PATH = "files/certificates/truststore.jks";
-	private static final String TRUSTSTORE_PASSWORD = "changeit";
-
 	static {
-		System.setProperty("jsse.enableSNIExtension", "false");
-		System.setProperty("javax.net.ssl.trustStore", TRUSTSTORE_PATH);
-		System.setProperty("javax.net.ssl.trustStorePassword", TRUSTSTORE_PASSWORD);
 		SLF4JBridgeHandler.removeHandlersForRootLogger();
 		SLF4JBridgeHandler.install();
 	}
@@ -81,6 +76,10 @@ public class ReportPortalTest {
 
 	private static final String KEYSTORE_PATH = "files/certificates/keystore.jks";
 	private static final String KEYSTORE_PASSWORD = "keystorePassword";
+	private static final String TRUSTSTORE_DEFAULT_PASSWORD_PATH = "files/certificates/truststore.jks";
+	private static final String TRUSTSTORE_PATH = "files/certificates/truststore_password.jks";
+	private static final String TRUSTSTORE_DEFAULT_PASSWORD = "changeit";
+	private static final String TRUSTSTORE_PASSWORD = "truststorePassword";
 
 	private static final String INVALID_KEYSTORE_PATH = "invalid/path/to/keystore.jks";
 	private static final String INVALID_KEYSTORE_PASSWORD = "invalidPassword";
@@ -235,15 +234,22 @@ public class ReportPortalTest {
 	private SSLServerSocket createSSLServerSocket(int port) throws Exception {
 		System.setProperty("javax.net.debug", "ssl,handshake");
 
-		KeyStore keyStore = KeyStore.getInstance("JKS");
-		keyStore.load(getClass().getResourceAsStream(KEYSTORE_PATH), KEYSTORE_PASSWORD.toCharArray());
+		KeyStore ks = KeyStore.getInstance("JKS");
+		InputStream ksis = Thread.currentThread().getContextClassLoader().getResourceAsStream(KEYSTORE_PATH);
+		ks.load(ksis, KEYSTORE_PASSWORD.toCharArray());
 
-		KeyManagerFactory kmf = KeyManagerFactory.getInstance(KeyManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm())
+		KeyManagerFactory kmf = KeyManagerFactory.getInstance(KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm())
 				.getAlgorithm());
-		kmf.init(keyStore, KEYSTORE_PASSWORD.toCharArray());
+		kmf.init(ks, KEYSTORE_PASSWORD.toCharArray());
+
+		KeyStore ts = KeyStore.getInstance("JKS");
+		InputStream tsis = Thread.currentThread().getContextClassLoader().getResourceAsStream(TRUSTSTORE_DEFAULT_PASSWORD_PATH);
+		ts.load(tsis, TRUSTSTORE_DEFAULT_PASSWORD.toCharArray());
+		TrustManagerFactory tmf = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+		tmf.init(ts);
 
 		SSLContext sslContext = SSLContext.getInstance("TLS");
-		sslContext.init(kmf.getKeyManagers(), null, null);
+		sslContext.init(kmf.getKeyManagers(), tmf.getTrustManagers(), new SecureRandom());
 
 		SSLServerSocketFactory sslServerSocketFactory = sslContext.getServerSocketFactory();
 		String[] enabledCipherSuites = sslServerSocketFactory.getSupportedCipherSuites();
@@ -412,6 +418,7 @@ public class ReportPortalTest {
 	public void verify_launch_uuid_creation_skip_parameter_handling() {
 		simulateStartLaunchResponse(rpClient);
 		simulateStartTestItemResponse(rpClient);
+		simulateFinishLaunchResponse(rpClient);
 
 		String launchUuid = "test-launch-uuid";
 		ListenerParameters listenerParameters = TestUtils.standardParameters();
@@ -451,9 +458,21 @@ public class ReportPortalTest {
 		}
 	}
 
-	@Test
-	@Disabled("FIXME: Unable to generate self-signed certificate or setup SSL correctly: SSLHandshakeException: no cipher suites in common")
-	public void verify_https_parameters_work_with_self_signed_certificate() throws Exception {
+	public static Iterable<Object[]> keystore_values() {
+		return Arrays.asList(
+				new Object[] { KEYSTORE_PATH, KEYSTORE_PASSWORD, TRUSTSTORE_DEFAULT_PASSWORD_PATH, TRUSTSTORE_DEFAULT_PASSWORD },
+				new Object[] { null, null, TRUSTSTORE_DEFAULT_PASSWORD_PATH, TRUSTSTORE_DEFAULT_PASSWORD },
+				new Object[] { KEYSTORE_PATH, KEYSTORE_PASSWORD, TRUSTSTORE_DEFAULT_PASSWORD_PATH, null },
+				new Object[] { KEYSTORE_PATH, KEYSTORE_PASSWORD, TRUSTSTORE_DEFAULT_PASSWORD_PATH, TRUSTSTORE_DEFAULT_PASSWORD },
+				new Object[] { KEYSTORE_PATH, KEYSTORE_PASSWORD, TRUSTSTORE_PATH, null },
+				new Object[] { KEYSTORE_PATH, KEYSTORE_PASSWORD, TRUSTSTORE_PATH, TRUSTSTORE_PASSWORD }
+		);
+	}
+
+	@ParameterizedTest
+	@MethodSource("keystore_values")
+	public void verify_https_parameters_work_with_self_signed_certificate(String keystorePath, String keystorePassword,
+			String truststorePath, String truststorePassword) throws Exception {
 		ServerSocket ss = SocketUtils.getServerSocketOnFreePort();
 		int port = ss.getLocalPort();
 		ss.close();
@@ -469,8 +488,10 @@ public class ReportPortalTest {
 
 		ListenerParameters parameters = TestUtils.standardParameters();
 		parameters.setBaseUrl(baseUrl);
-		parameters.setKeystore(KEYSTORE_PATH);
-		parameters.setKeystorePassword(KEYSTORE_PASSWORD);
+		parameters.setKeystore(keystorePath);
+		parameters.setKeystorePassword(keystorePassword);
+		parameters.setTruststore(truststorePath);
+		parameters.setTruststorePassword(truststorePassword);
 
 		ExecutorService clientExecutor = Executors.newSingleThreadExecutor();
 		ReportPortalClient rpClient = ReportPortal.builder().buildClient(ReportPortalClient.class, parameters, clientExecutor);
@@ -498,12 +519,22 @@ public class ReportPortalTest {
 		);
 	}
 
+	@Test
+	public void verify_invalid_truststore_path() {
+		ListenerParameters parameters = TestUtils.standardParameters();
+		parameters.setBaseUrl("https://localhost:8443");
+		parameters.setTruststore(INVALID_KEYSTORE_PATH);
+		parameters.setTruststorePassword(TRUSTSTORE_DEFAULT_PASSWORD);
+
+		ExecutorService clientExecutor = Executors.newSingleThreadExecutor();
+		assertThrows(
+				InternalReportPortalClientException.class,
+				() -> ReportPortal.builder().buildClient(ReportPortalClient.class, parameters, clientExecutor)
+		);
+	}
 
 	public static Iterable<Object[]> invalid_keystore_passwords() {
-		return Arrays.asList(
-				new Object[] { INVALID_KEYSTORE_PASSWORD },
-				new Object[] { null }
-		);
+		return Arrays.asList(new Object[] { INVALID_KEYSTORE_PASSWORD }, new Object[] { null });
 	}
 
 	@ParameterizedTest
@@ -513,6 +544,20 @@ public class ReportPortalTest {
 		parameters.setBaseUrl("https://localhost:8443");
 		parameters.setKeystore(KEYSTORE_PATH);
 		parameters.setKeystorePassword(password);
+
+		ExecutorService clientExecutor = Executors.newSingleThreadExecutor();
+		assertThrows(
+				InternalReportPortalClientException.class,
+				() -> ReportPortal.builder().buildClient(ReportPortalClient.class, parameters, clientExecutor)
+		);
+	}
+
+	@Test
+	public void verify_invalid_truststore_password() {
+		ListenerParameters parameters = TestUtils.standardParameters();
+		parameters.setBaseUrl("https://localhost:8443");
+		parameters.setTruststore(TRUSTSTORE_DEFAULT_PASSWORD_PATH);
+		parameters.setTruststorePassword(INVALID_KEYSTORE_PASSWORD);
 
 		ExecutorService clientExecutor = Executors.newSingleThreadExecutor();
 		assertThrows(