+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [5.2.30]
+### Added
+- `rp.truststore.resource` and `rp.truststore.password` configuration parameters, by @HardNorth
+### Fixed
+- Manual SSL configuration for `OkHttpClient`, by @HardNorth
+### Changed
+- `LaunchLoggingContext.completed` is now public, by @HardNorth
+
 ## [5.2.29]
 ### Fixed
 - `NullPointerException` in case of no binary data in `ReportPortal.toSaveLogRQ` method, by @HardNorth